Fix potential wrong checksum issue in resumed backups.
To trigger the issue:

1. A file with an existing checksum in the previous backup is changed
2. The backup is interrupted _after_ the file change is transferred
   by rsync, but before the new checksum is stored
3. The file stays unchanged until the backup is later resumed and
   completed.
4. As the file is already transferred to the incomplete backup its
   checksum will not be outputed by rsync and the old checksum from
   the previos complete backup was used.

This issue was fixed by running an extra rsync in dry run mode
between old and new backup to get all changed files between the
backups and ensuring these files get their checksums calculated.
The files with checksums from rsync in the finishing rsync run
is not recalculated again although they are in the changed files
set to avoid doing uneccessary work.